iom/requirements.txt
Django==2.2.6
psycopg2==2.8.4

pillow==6.2.1
djangorestframework==3.10.3